style: change channel index composition
